jk need the 1/J
