+++ src/constitutive/solid/elastic_hyperviscous/almansi_hamel/mod.rs
@@ -74,13 +74,13 @@ impl<'a> Viscoelastic<'a> for AlmansiHamel<'a>
     fn calculate_cauchy_stress(&self, deformation_gradient: &DeformationGradient, deformation_gradient_rate: &DeformationGradientRate) -> CauchyStress
     {
         let identity = LeftCauchyGreenDeformation::identity();
-        let inverse_deformation_gradient = deformation_gradient.inverse();
+        let (inverse_deformation_gradient, jacobian) = deformation_gradient.inverse_and_determinant();
         let strain = (&identity * 1.0 - inverse_deformation_gradient.transpose() * &inverse_deformation_gradient) * 0.5;
         let (deviatoric_strain, strain_trace) = strain.deviatoric_and_trace();
         let velocity_gradient = deformation_gradient_rate * inverse_deformation_gradient;
         let strain_rate = (&velocity_gradient + velocity_gradient.transpose()) * 0.5;
         let (deviatoric_strain_rate, strain_rate_trace) = strain_rate.deviatoric_and_trace();
-        deviatoric_strain*(2.0*self.get_shear_modulus()) + deviatoric_strain_rate*(2.0*self.get_shear_viscosity()) + CauchyStress::identity()*(self.get_bulk_modulus()*strain_trace + self.get_bulk_viscosity()*strain_rate_trace)
+        deviatoric_strain*(2.0*self.get_shear_modulus()/jacobian) + deviatoric_strain_rate*(2.0*self.get_shear_viscosity()/jacobian) + CauchyStress::identity()*((self.get_bulk_modulus()*strain_trace + self.get_bulk_viscosity()*strain_rate_trace)/jacobian)
     }
     /// Calculates and returns the rate tangent stiffness associated with the Cauchy stress.
     ///
@@ -90,8 +90,24 @@ impl<'a> Viscoelastic<'a> for AlmansiHamel<'a>
     fn calculate_cauchy_rate_tangent_stiffness(&self, deformation_gradient: &DeformationGradient, _: &DeformationGradientRate) -> CauchyRateTangentStiffness
     {
         let identity = CauchyStress::identity();
-        let deformation_gradient_inverse_transpose = deformation_gradient.inverse_transpose();
-        let scaled_deformation_gradient_inverse_transpose = &deformation_gradient_inverse_transpose * self.get_shear_viscosity();
-        CauchyRateTangentStiffness::dyad_ik_jl(&identity, &scaled_deformation_gradient_inverse_transpose) + CauchyRateTangentStiffness::dyad_il_jk(&scaled_deformation_gradient_inverse_transpose, &identity) + CauchyRateTangentStiffness::dyad_ij_kl(&(identity*(self.get_bulk_viscosity() - 2.0/3.0*self.get_shear_viscosity())), &deformation_gradient_inverse_transpose)
+        let (deformation_gradient_inverse_transpose, jacobian) = deformation_gradient.inverse_transpose_and_determinant();
+        let scaled_deformation_gradient_inverse_transpose = &deformation_gradient_inverse_transpose * self.get_shear_viscosity()/jacobian;
+        CauchyRateTangentStiffness::dyad_ik_jl(&identity, &scaled_deformation_gradient_inverse_transpose) + CauchyRateTangentStiffness::dyad_il_jk(&scaled_deformation_gradient_inverse_transpose, &identity) + CauchyRateTangentStiffness::dyad_ij_kl(&(identity*((self.get_bulk_viscosity() - 2.0/3.0*self.get_shear_viscosity())/jacobian)), &deformation_gradient_inverse_transpose)
+    }
+}
+
+/// Elastic-hyperviscous constitutive model implementation of the Almansi-Hamel elastic-hyperviscous constitutive model.
+impl<'a> ElasticHyperviscous<'a> for AlmansiHamel<'a>
+{
+    /// Calculates and returns the viscous dissipation.
+    ///
+    /// ```math
+    /// \phi(\mathbf{F},\dot{\mathbf{F}}) = \eta\,\mathrm{tr}(\mathbf{D}^2) + \frac{1}{2}\left(\zeta - \frac{2}{3}\,\eta\right)\mathrm{tr}(\mathbf{D})^2
+    /// ```
+    fn calculate_viscous_dissipation(&self, deformation_gradient: &DeformationGradient, deformation_gradient_rate: &DeformationGradientRate) -> Scalar
+    {
+        let velocity_gradient = deformation_gradient_rate * deformation_gradient.inverse();
+        let strain_rate = (&velocity_gradient + velocity_gradient.transpose()) * 0.5;
+        self.get_shear_viscosity()*strain_rate.squared_trace() + 0.5*(self.get_bulk_viscosity() - 2.0/3.0*self.get_shear_viscosity())*strain_rate.trace().powi(2)
     }
 }
