+++ src/constitutive/solid/hyperelastic/saint_venant_kirchoff/mod.rs
@@ -79,7 +79,7 @@ impl<'a> Hyperelastic<'a> for SaintVenantKirchoff<'a>
     /// Calculates and returns the Helmholtz free energy density.
     ///
     /// ```math
-    /// a(\mathbf{F}) = \mu\,\mathrm{tr}(\mathbf{E}^2) + \left(\kappa - \frac{2}{3}\,\mu\right)\mathrm{tr}(\mathbf{E})^2
+    /// a(\mathbf{F}) = \mu\,\mathrm{tr}(\mathbf{E}^2) + \frac{1}{2}\left(\kappa - \frac{2}{3}\,\mu\right)\mathrm{tr}(\mathbf{E})^2
     /// ```
     fn calculate_helmholtz_free_energy_density(&self, deformation_gradient: &DeformationGradient) -> Scalar
     {
