+++ src/constitutive/solid/thermoelastic/almansi_hamel/mod.rs
@@ -56,27 +56,29 @@ impl<'a> Thermoelastic<'a> for AlmansiHamel<'a>
     /// Calculates and returns the Cauchy stress.
     ///
     /// ```math
-    /// \boldsymbol{\sigma}(\mathbf{F}, T) = 2\mu\mathbf{e}' + \kappa\,\mathrm{tr}(\mathbf{e})\mathbf{1} - 3\alpha\kappa(T - T_\mathrm{ref})\mathbf{1}
+    /// \boldsymbol{\sigma}(\mathbf{F}, T) = \frac{2\mu}{J}\,\mathbf{e}' + \frac{\kappa}{J}\,\mathrm{tr}(\mathbf{e})\mathbf{1} - \frac{3\alpha\kappa}{J}(T - T_\mathrm{ref})\mathbf{1}
     /// ```
     fn calculate_cauchy_stress(&self, deformation_gradient: &DeformationGradient, temperature: &Scalar) -> CauchyStress
     {
         let identity = LeftCauchyGreenDeformation::identity();
-        let inverse_deformation_gradient = deformation_gradient.inverse();
+        let (inverse_deformation_gradient, jacobian) = deformation_gradient.inverse_and_determinant();
         let strain = (&identity * 1.0 - inverse_deformation_gradient.transpose() * &inverse_deformation_gradient) * 0.5;
         let (deviatoric_strain, strain_trace) = strain.deviatoric_and_trace();
-        deviatoric_strain * (2.0 * self.get_shear_modulus()) + identity * (self.get_bulk_modulus() * (strain_trace - 3.0*self.get_coefficient_of_thermal_expansion()*(temperature - self.get_reference_temperature())))
+        deviatoric_strain * (2.0 * self.get_shear_modulus() / jacobian) + identity * (self.get_bulk_modulus() / jacobian * (strain_trace - 3.0*self.get_coefficient_of_thermal_expansion()*(temperature - self.get_reference_temperature())))
     }
     /// Calculates and returns the tangent stiffness associated with the Cauchy stress.
     ///
     /// ```math
-    /// \mathcal{T}_{ijkL}(\mathbf{F}) = \mu B_{jk}^{-1}F_{iL}^{-T} + \mu B_{ik}^{-1}F_{jL}^{-T} + \left(\kappa - \frac{2}{3}\,\mu\right)\delta_{ij}B_{km}^{-1}F_{mL}^{-T}
+    /// \mathcal{T}_{ijkL}(\mathbf{F}, T) = \frac{\mu}{J}\left[B_{jk}^{-1}F_{iL}^{-T} + B_{ik}^{-1}F_{jL}^{-T} - \frac{2}{3}\,\delta_{ij}B_{km}^{-1}F_{mL}^{-T} - 2e_{ij}'F_{kL}^{-T}\right] + \frac{\kappa}{J}\left\{\delta_{ij}B_{km}^{-1}F_{mL}^{-T} - \Big[\mathrm{tr}(\mathbf{e}) - 3\alpha(T - T_\mathrm{ref})\Big]\delta_{ij}F_{kL}^{-T}\right\}
     /// ```
-    fn calculate_cauchy_tangent_stiffness(&self, deformation_gradient: &DeformationGradient, _: &Scalar) -> CauchyTangentStiffness
+    fn calculate_cauchy_tangent_stiffness(&self, deformation_gradient: &DeformationGradient, temperature: &Scalar) -> CauchyTangentStiffness
     {
         let identity = LeftCauchyGreenDeformation::identity();
-        let inverse_transpose_deformation_gradient = deformation_gradient.inverse_transpose();
+        let (inverse_transpose_deformation_gradient, jacobian) = deformation_gradient.inverse_transpose_and_determinant();
         let inverse_left_cauchy_green_deformation = &inverse_transpose_deformation_gradient * inverse_transpose_deformation_gradient.transpose();
-        (CauchyTangentStiffness::dyad_il_jk(&inverse_transpose_deformation_gradient, &inverse_left_cauchy_green_deformation) + CauchyTangentStiffness::dyad_ik_jl(&inverse_left_cauchy_green_deformation, &inverse_transpose_deformation_gradient)) * self.get_shear_modulus() + CauchyTangentStiffness::dyad_ij_kl(&identity, &(inverse_left_cauchy_green_deformation * &inverse_transpose_deformation_gradient*(self.get_bulk_modulus() - self.get_shear_modulus() * 2.0 / 3.0)))
+        let strain = (&identity * 1.0 - &inverse_left_cauchy_green_deformation) * 0.5;
+        let (deviatoric_strain, strain_trace) = strain.deviatoric_and_trace();
+        (CauchyTangentStiffness::dyad_il_jk(&inverse_transpose_deformation_gradient, &inverse_left_cauchy_green_deformation) + CauchyTangentStiffness::dyad_ik_jl(&inverse_left_cauchy_green_deformation, &inverse_transpose_deformation_gradient)) * (self.get_shear_modulus() / jacobian)+ CauchyTangentStiffness::dyad_ij_kl(&identity,&(inverse_left_cauchy_green_deformation * &inverse_transpose_deformation_gradient*((self.get_bulk_modulus() - self.get_shear_modulus() * 2.0 / 3.0) / jacobian))) - CauchyTangentStiffness::dyad_ij_kl(&(deviatoric_strain * (2.0 * self.get_shear_modulus() / jacobian) + identity * (self.get_bulk_modulus() / jacobian * (strain_trace - 3.0*self.get_coefficient_of_thermal_expansion()*(temperature - self.get_reference_temperature())))), &inverse_transpose_deformation_gradient)
     }
     fn get_coefficient_of_thermal_expansion(&self) -> &Scalar
     {
