+++ src/constitutive/solid/viscoelastic/mod.rs
@@ -11,10 +11,6 @@
 #[cfg(test)]
 pub mod test;
 
-mod almansi_hamel;
-
-pub use almansi_hamel::AlmansiHamel;
-
 use super::
 {
     *, super::fluid::viscous::Viscous
